Split out loginUser
